# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'E-commerce Modular API'
  description: 'Tài liệu API hệ thống E-commerce (Sử dụng PHP 8 Attributes)'
  contact:
    email: admin@system.com
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API Server'
paths:
...
  /admin/users:
    get:
      tags:
        - 'Users (Admin & Self)'
      summary: 'Xem danh sách người dùng (Admin)'
      operationId: ebdb37e0aca7226f6630d2b6f1de6946
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Success
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
    post:
      tags:
        - 'Users (Admin & Self)'
      summary: 'Tạo người dùng mới (Admin)'
      operationId: 88de0372c28d80c0be353a8bcce497e7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                roles:
                  type: array
                  items:
                    type: string
              type: object
      responses:
        201:
          description: Created
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
  '/admin/users/{uuid}':
    get:
      tags:
        - 'Users (Admin & Self)'
      summary: 'Xem chi tiết người dùng'
      operationId: 0c8ff92257b9f5d7e945b3a8ee6edad6
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
        403:
          description: 'Forbidden (Not Owner/Admin)'
        404:
          description: 'Not Found'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - 'Users (Admin & Self)'
      summary: 'Cập nhật thông tin'
      operationId: 2d7ab603a4edd1e1187a11eefe3bb856
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
              type: object
      responses:
        200:
          description: Updated
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Users (Admin & Self)'
      summary: 'Xóa người dùng (Admin)'
      operationId: 13d69b0c1b0fb851225843080eac0c32
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
  /profile:
    get:
      tags:
        - 'Users (Admin & Self)'
      summary: 'Xem thông tin người dùng hiện tại'
      operationId: 02215cf763020c25fc4b65d33aa9cb9f
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                  data:
                    properties:
                      uuid:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      roles:
                        type: array
                        items:
                          type: string
                    type: object
                type: object
        404:
          description: 'User not found'
      security:
        -
          bearerAuth: []
  /admin/warehouses:
    get:
      tags:
        - Warehouses
      summary: 'Xem danh sách Kho hàng (Admin)'
      operationId: 6320efc130774dac5aa7c13dafd0ae2b
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Success
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Warehouses
      summary: 'Tạo Kho hàng mới (Admin)'
      operationId: 1a7c4e3a6a3f36dc8ca5a75d54dff1db
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - code
              properties:
                name:
                  type: string
                  example: 'Kho Trung Tâm'
                code:
                  type: string
                  example: NB01
                address:
                  type: string
              type: object
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  '/admin/warehouses/{uuid}':
    get:
      tags:
        - Warehouses
      summary: 'Xem chi tiết Kho hàng'
      operationId: 6e939c49cfde0998a14c14a181295f74
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
        404:
          description: 'Not found'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Warehouses
      summary: 'Cập nhật Kho hàng'
      operationId: 9ff30ab1a0694145ef8eb264c224255f
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                address:
                  type: string
              type: object
      responses:
        200:
          description: Updated
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Warehouses
      summary: 'Xóa Kho hàng'
      operationId: d20e7ee47b740f732b6fbdb123213f44
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []